+++ goodmorning.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+
+from scenes import *
+
+# Bring up the Accents
+hueaccent(on=True)
+
+# Living Room Normal (except downlights)
+
+# Bedroom Normal (half-ish, except downlights)
+
+# All downlights on full cool.
+kdl()
+